+++ .idea/misc.xml
@@ -64,6 +64,7 @@
         <entry key="../../../../../layout/compose-model-1651069174293.xml" value="2.0" />
         <entry key="../../../../../layout/compose-model-1651070610878.xml" value="0.20059121621621623" />
         <entry key="../../../../../layout/compose-model-1651070620216.xml" value="0.7885057471264367" />
+        <entry key="../../../../../layout/compose-model-1651160350021.xml" value="0.1787037037037037" />
         <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.131" />
         <entry key="app/src/main/res/drawable/ic_add.xml" value="0.182" />
         <entry key="app/src/main/res/drawable/ic_bar_daily.xml" value="0.131" />
