+++ app/src/main/java/com/mightyhedgehog/doplanner/data/local/datastore/ThemeDataStore.kt
@@ -6,7 +6,7 @@ import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.datastore.preferences.preferencesDataStore
-import com.mightyhedgehog.doplanner.ui.theme.style.DoPlannerStyle
+import com.mightyhedgehog.doplanner.ui.theme.DoPlannerStyle
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
