+++ app/src/main/java/com/mightyhedgehog/doplanner/domain/model/settings/SettingsBundle.kt
@@ -1,6 +1,6 @@
 package com.mightyhedgehog.doplanner.domain.model.settings
 
-import com.mightyhedgehog.doplanner.ui.theme.style.DoPlannerStyle
+import com.mightyhedgehog.doplanner.ui.theme.DoPlannerStyle
 
 data class SettingsBundle(
     val style: DoPlannerStyle
