+++ app/src/main/java/com/mightyhedgehog/doplanner/ui/theme/DoPlannerCommon.kt
@@ -15,7 +15,7 @@ data class DoPlannerColors(
     val backgroundWhite: Color,
     val grey: Color,
     val mainColor: Color,
-    val secondOrange: Color,
+    val secondaryColor: Color,
     val redBackground: Color,
     val blueBackground: Color,
     val greenBackground: Color,
