+++ app/src/main/java/com/mightyhedgehog/doplanner/ui/theme/DoPlannerStyle.kt
@@ -1,0 +1,5 @@
+package com.mightyhedgehog.doplanner.ui.theme
+
+enum class DoPlannerStyle {
+    ORANGE, BLUE, PINK, PURPLE, GREEN
+}
