--- app/src/main/java/com/mightyhedgehog/doplanner/ui/theme/style/DoPlannerStyle.kt
@@ -1,12 +1,0 @@
-package com.mightyhedgehog.doplanner.ui.theme.style
-
-import kotlinx.coroutines.flow.StateFlow
-
-enum class DoPlannerStyle {
-    ORANGE, BLUE, PINK, PURPLE, GREEN
-}
-
-interface UserSettings {
-    val styleStream: StateFlow<DoPlannerStyle>
-    val style: DoPlannerStyle
-}
